Add checklist icon
The list-bullet and numbered-list icons are different both in their
naming scheme and icon style. I chose to base this icon on numbered-list
because it can not be made smaller while list-bullet can be made bigger.
Perhaps I can make a bullet-list in the future that is similar to
checklist and numbered-list.